requirements.txt
requests>=2.31.0
numpy>=1.24.0
faiss-cpu>=1.9.0
pymilvus>=2.3.0
python-dotenv>=1.0.0
tqdm>=4.66.0
PyMuPDF>=1.23.0
Pillow>=10.0.0
PyYAML>=6.0
docling>=2.0.0